refactor: split small web and snapshot boundaries
